requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
seaborn
numpy==1.23.5
pandas==1.5.3
scipy==1.10.0
scikit-learn==1.2.1
xgboost

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-bigquery
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor

# data cleaning
fancyimpute

#stat models
statsmodels

#sklearn
scikit-learn

nbresult
colorama
ipdb
ipykernel
yapf
tensorflow
matplotlib
pygeohash
pytest
db-dtypes
pyarrow
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

# prevent bq 3+ db-dtypes package error when running training from bq

# API
fastapi
pytz
uvicorn
# tests
httpx
pytest-asyncio
